requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
cryptography==38.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.3
django-allauth==0.51.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
django-taggit==3.0.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.1
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
